+++ src/main/java/step3/Main.java
@@ -1,5 +1,9 @@
 package step3;
 
+import step3.racing.CarRacingGame;
+import step3.view.InputView;
+import step3.view.ResultView;
+
 /**
  * @author : 0giri
  * @since : 2023/04/16
