+++ src/main/java/step3/racing/Car.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.racing;
 
 /**
  * @author : 0giri
