+++ src/main/java/step3/racing/CarActionDecider.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.racing;
 
 import java.util.function.Consumer;
