+++ src/main/java/step3/racing/Cars.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.racing;
 
 import java.util.ArrayList;
 import java.util.Collections;
