+++ src/main/java/step3/racing/RacingResult.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.racing;
 
 import java.util.List;
