+++ src/test/java/step3/CarRacingGameTest.java
@@ -2,6 +2,7 @@ package step3;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import step3.racing.CarRacingGame;
 
 import static org.assertj.core.api.Assertions.*;
