+++ src/test/java/step3/CarTest.java
@@ -2,6 +2,7 @@ package step3;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import step3.racing.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
