+++ src/test/java/step3/RandomGeneratorTest.java
@@ -2,6 +2,7 @@ package step3;
 
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
+import step3.util.RandomGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
